+++ practice-app/server/src/server.js
@@ -1,7 +1,7 @@
 const http = require("http");
 const mongoose = require('mongoose');
-const jwkToPem = require("jwk-to-pem");
-const axios = require("axios");
+//const axios = require("axios");
+const jwks = require("./services/jwkeys")
 require("dotenv").config();
 
 const app = require("./app");
@@ -10,7 +10,7 @@ const PORT = process.env.API_PORT;
 const server = http.createServer(app);
 const DB = process.env.DB;
 const AUTH0_DOMAIN = process.env.auth0_domain
-var KEYS =[];
+// var KEYS =[];
 
 const connectDB = async () => {
     await mongoose
@@ -38,22 +38,28 @@ const connectDB = async () => {
 
 };
 
-const getJwks = async () => {
-    const url = `https://${AUTH0_DOMAIN}/.well-known/jwks.json`;
-    const response = (await axios.get(url)).data;
-    //const pem = jwkToPem()
-    //console.log(response);
-    response.keys.forEach(function(e,i,a) {
-        //console.log("--->>", e);
-        const pem = jwkToPem(e)
-        KEYS.push(pem);
-    });
-    KEYS.forEach(function(e,i,a) {
-        //console.log("--->>", e);
-    });
-};
+// const getJwks = async () => {
+//     const url = `https://${AUTH0_DOMAIN}/.well-known/jwks.json`;
+//     const response = (await axios.get(url)).data;
+//     //const pem = jwkToPem()
+//     //console.log(response);
+//     response.keys.forEach(function(e,i,a) {
+//         //console.log("--->>", e);
+//         const pem = jwkToPem(e)
+//         KEYS.push(pem);
+//     });
+//     KEYS.forEach(function(e,i,a) {
+//         //console.log("--->>", e);
+//     });
+// };
 
-getJwks();
+// const getKeys = () => {
+//     console.log(KEYS);
+//     return KEYS;
+// }
+
+//getJwks();
+jwks.getJwks();
 
 async function startServer() {
   server.listen(PORT, () => {
